+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - Added support for `DataFrame.unstack` and `Series.unstack`.
 - Added support for `DataFrame.asfreq` and `Series.asfreq`.
 - Added support for `Series.dt.is_month_start` and `Series.dt.is_month_end`.
+- Added support for `Series.dt.is_year_start` and `Series.dt.is_year_end`.
 
 #### Improvements
 - Removed the public preview warning message upon importing Snowpark pandas.